+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 Changes:
 
+  * Added consumables with LOT numbers to Illumina runs.
 
 
 # 2.2.0